backendPython/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
dnspython==2.2.1
joblib==1.1.0
numpy==1.23.1
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pymongo==4.2.0
pyparsing==3.0.9
pytesseract==0.3.9
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==1.1.2
scipy==1.9.0
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
